boss hp from int to float, cinemachine package

+++ Assets/Scripts/BossController.cs
@@ -40,7 +40,7 @@ public class BossController : MonoBehaviour
 
         RegenerateShootPoints();
         ActNextOnSequence();
-        bossHealthBar.SetBossMaxHealth((int)Hp);
+        bossHealthBar.SetBossMaxHealth(Hp);
         bossHealthBar.SetHealthBarActive();
     }
